requirements.txt
# Flask Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# Environment Variables
python-dotenv==1.0.0

# File Type Detection
python-magic==0.4.27

# Email Parsing & HTML Processing (Phase 2)
html2text==2020.1.16
email-validator==2.1.0

# Database
# sqlite3 (built into Python)

# Security
bcrypt==4.1.2
PyJWT==2.8.0

# Development & Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==6.2.1

# Production Server
gunicorn==21.2.0

# AI Integration (Phase 3) - Active
openai>=1.35.0  # Updated for compatibility
jsonschema==4.20.0
Flask-Limiter==3.5.0

# Future URL Analysis (Phase 4)
# requests==2.31.0
# urllib3==2.1.0

# Future Database Migration (Phase 6)
# psycopg2-binary==2.9.7
# SQLAlchemy==2.0.23
# Flask-Migrate==4.0.5

# Demo Generation Tools
playwright==1.40.0
pillow==10.1.0

# Future Caching (Phase 6)
# redis==5.0.1
# Flask-Caching==2.1.0